// src/module/entities/model/OrderItem.ts
export class OrderItem {
  identifier: number;
  positionItem: number;
  percentagePis: number;
  percentageCofins: number;
  order: Order;
  itemNumber: number;
  entryDate: string;
  deliveryDate: string;
  product: Product;
  quantity: number;
  stockUnit: {
    unit: string;
    description: string;
    abbreviation: string;
    lastChangeDate: string;
    lastChangeTime: number;
  };
  weight: number;
  packing: {
    code: number;
    description: string;
    abbreviation: string;
    lastChangeDate: string;
    lastChangeTime: number;
  };
  priceList: number;
  price: number;
  standardPrice: number;
  unitCost: number;
  productReference: string;
  productDescription: string;
  purschaseOrder: string;
  itemPurchaseOrder: number;
  externalIdOrder: number;
  combination: number;
  originOfGoods: number;
  movimentedBatch: {
    code: number;
    batchNumber: string;
    supplierBatchNumber: string;
    manufacturingDate: string;
    expirationDate: string;
    availabilityDate: string;
    lastChangeDate: string;
    lastChangeTime: number;
  };
  stockLocation: {
    code: number;
    description: string;
    abbreviation: string;
    lastChangeDate: string;
    lastChangeTime: number;
  };
  firstPercentageDiscount: number;
  secondPercentageDiscount: number;
  thirdPercentageDiscount: number;
  fourthPercentageDiscount: number;
  fifthPercentageDiscount: number;
  percentageIncrease: number;
  icmtCst: number;
  percentageIcms: number;
  percentageIpi: number;
  taxSituationPis: number;
  taxSituationCofins: number;
  taxSituationIpi: number;
  simplesNacionalOperationCode: number;
  grossAmount: number;
  discountAmount: number;
  totalNetAmount: number;
  issPercentageRetained: number;
  perIcmsSubstituteBaseReduction: number;
  freightAmount: number;
  insuranceAmount: number;
  ancillaryExpensesAmount: number;
  individualPackageType: number;
  issPercentageNotRetained: number;
  numberOfLabels: number;
  itemIdObservation: Observation;
  productionIdObservation: Observation;
  lastChangeDate: string;
  lastChangeTime: number;
  cancellationDate: string;
  cancellationTime: number;
  itemPurchaseOrderSiger: number;
}